preparation by splitting tests
