+++ Makefile
@@ -66,7 +66,7 @@ test-coverage-with-graphdb:  ## runs the standard test-suite for all available i
 check:  ## performs linting on the python code
 	@poetry run black --check --diff .
 	@poetry run isort --check --diff .
-	@poetry run flake8 . --exclude ${FLAKE8_EXCLUDE} --ignore=E501,E201,E202,W503
+	@poetry run flake8 . --exclude ${FLAKE8_EXCLUDE}
 
 lint-fix:  ## fixes code according to the lint suggestions
 	@poetry run black .
